requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
telethon==1.34.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.1
python-multipart==0.0.6
aiofiles==23.2.1
cryptography==42.0.0
qrcode==7.4.2
Pillow==10.2.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
